(refactor) Refactor feed components to be updated with use of new workout model
